Create README.md
In theory, Siamese networks are a kind of neural architecture that measures how different two images are. The inputs for the Siamese Network are a pair of images. The images are positive if they are similar and their target value will be set to 0. On the other hand, the images are negative if they are different and their target value will be set to 1.